refactor(scaffold): extract the repo-scaffold engine into internal/scaffold
Moves the scaffold engine shared by `new` and `boot` out of package cmd
into a pure internal/scaffold package (inputs: project root, schema,
framework source dir; effect: scaffold files written):
- engine.go: FindTable (manifest-driven), FindTableInSchemas
  (schema-list-driven, from boot.go), RepoForTable, CustomRepo
- bridgeyml.go: bridge.yml template + BridgeYMLForTable
- queries.go: Queries + its private spec/slug helpers
- ancestry.go: ParentInfo, Ancestry, DetectAncestry

cmd keeps flag parsing, loadProject, and dispatch; new.go/boot.go now
call the scaffold package. The five scaffold tests moved verbatim from
cmd/new_test.go to internal/scaffold with assertions unchanged.
Function bodies, templates, progress output, and error text are
byte-identical copies — only package and export renames.

Prepares P1.3 (codegen lift into the framework): the scaffold engine is
now a self-contained package the in-framework tool can consume.

Matrix green (S0-S10 x all stages), determinism intact.
